@@ -1,10 +1,10 @@
 name: bloc_rest_api
 description: A generic bloc package to implement REST APIs easily in flutter.
-version: 0.4.0+1
+version: 1.0.0-nullsafety.0
 homepage: https://github.com/muhammadidrees/bloc_rest_api
 
 environment: 
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
@@ -16,6 +16,7 @@ dependencies:
   http: ^0.13.3
   hydrated_bloc: ^7.0.1
   meta: ^1.3.0
+  build_runner: ^2.0.6
 
 dev_dependencies:
   flutter_test:
