@@ -2,12 +2,10 @@ part of 'cubits.dart';
 
 class HydratedRequestCubit<T> extends RequestCubit<T> with HydratedMixin {
   HydratedRequestCubit({
-    @required T Function(dynamic json) fromMap,
-    @required this.toMap,
-    http.Client httpClient,
-  })  : assert(fromMap != null, 'FromMap function cannot be null'),
-        assert(toMap != null, 'toMap function cannot be null'),
-        super(
+    required T Function(dynamic json) fromMap,
+    required this.toMap,
+    http.Client? httpClient,
+  }) : super(
           fromMap: fromMap,
           httpClient: httpClient,
         ) {
@@ -23,13 +21,13 @@ class HydratedRequestCubit<T> extends RequestCubit<T> with HydratedMixin {
         status: json['status'] == null
             ? null
             : enumFromString(json['status'].toString()),
-        model: json['model'] == null ? null : fromMap(json['model']),
+        model: json['model'] == null ? null : fromMap!(json['model']),
       );
 
   @override
   Map<String, dynamic> toJson(RequestState<T> state) => {
-        'status': state?.status?.toString() ?? RequestState<T>.empty(),
-        'model': state?.model == null ? null : toMap(state.model),
+        'status': state.status?.toString() ?? RequestState<T>.empty(),
+        'model': state.model == null ? null : toMap(state.model!),
       };
 
   @override
