@@ -8,14 +8,11 @@ import 'package:http/http.dart' as http;
 import 'package:bloc_test/bloc_test.dart';
 
 import '../models/models.dart';
-
-// Create a MockClient using the Mock class provided by the Mockito package.
-// Create new instances of this class in each test.
-class MockClient extends Mock implements http.Client {}
+import '../usage_test.mocks.dart';
 
 void main() {
-  http.Client client;
-  RequestCubit<PostModel> cubit;
+  http.Client? client;
+  late RequestCubit<PostModel> cubit;
 
   setUp(() {
     client = MockClient();
@@ -26,7 +23,7 @@ void main() {
   });
 
   tearDown(() {
-    client.close();
+    client!.close();
     cubit.close();
   });
 
@@ -34,8 +31,8 @@ void main() {
     blocTest(
       'emits [loading, success] on successful request',
       build: () => cubit,
-      act: (bloc) {
-        when(client.post(
+      act: (RequestCubit<PostModel> bloc) {
+        when(client!.post(
           Uri.parse('https://jsonplaceholder.typicode.com/posts/1'),
           body: {'post': 1},
         )).thenAnswer(
@@ -62,8 +59,8 @@ void main() {
     blocTest(
       'emits [loading, failure] on request fail',
       build: () => cubit,
-      act: (bloc) {
-        when(client.post(
+      act: (RequestCubit<PostModel> bloc) {
+        when(client!.post(
           Uri.parse('https://jsonplaceholder.typicode.com/posts/1'),
           body: {'post': 1},
         )).thenAnswer((_) async => http.Response('NotFound', 404));
@@ -88,8 +85,8 @@ void main() {
       'emits [loading, failure] on timeout fail',
       build: () => cubit,
       wait: Duration(milliseconds: 400),
-      act: (bloc) {
-        when(client.post(
+      act: (RequestCubit<PostModel> bloc) {
+        when(client!.post(
           Uri.parse('https://jsonplaceholder.typicode.com/posts/1'),
           body: {'post': 1},
         )).thenAnswer(
@@ -120,8 +117,8 @@ void main() {
       'emits [loading, success] when response within timeout',
       build: () => cubit,
       wait: Duration(milliseconds: 200),
-      act: (bloc) {
-        when(client.post(
+      act: (RequestCubit<PostModel> bloc) {
+        when(client!.post(
                 Uri.parse('https://jsonplaceholder.typicode.com/posts/1')))
             .thenAnswer(
           (_) async {
@@ -152,8 +149,8 @@ void main() {
         status: RequestStatus.success,
         model: PostModel(userId: 1, id: 1),
       ),
-      act: (bloc) {
-        when(client.post(
+      act: (RequestCubit<PostModel> bloc) {
+        when(client!.post(
                 Uri.parse('https://jsonplaceholder.typicode.com/posts/1')))
             .thenAnswer((_) async => http.Response('BadRequestException', 400));
 
