@@ -1,21 +1,19 @@
 import 'package:bloc_rest_api/bloc_rest_api.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:http/http.dart' as http;
 import 'package:bloc_test/bloc_test.dart';
 
 import 'models/models.dart';
+import 'usage_test.mocks.dart';
 
-// Create a MockClient using the Mock class provided by the Mockito package.
-// Create new instances of this class in each test.
-class MockClient extends Mock implements http.Client {}
-
+@GenerateMocks([http.Client])
 void main() {
   group(
     'check fromMap assertion',
     () {
-      final client = MockClient();
-
+      http.Client client = MockClient();
       when(client
               .get(Uri.parse('https://jsonplaceholder.typicode.com/posts/1')))
           .thenAnswer(
@@ -37,7 +35,7 @@ void main() {
       blocTest(
         'pass if fromMap is passed with cubit instead of function',
         build: () => cubitWithFromMap,
-        act: (bloc) => bloc.getRequest(
+        act: (RequestCubit<PostModel> bloc) => bloc.getRequest(
           baseUrl: 'https://jsonplaceholder.typicode.com/',
           handle: 'posts/1',
         ),
@@ -46,7 +44,7 @@ void main() {
       blocTest(
         'pass if fromMap is passed with the function instead with cubit',
         build: () => cubitWithoutFromMap,
-        act: (bloc) => bloc.getRequest(
+        act: (RequestCubit<PostModel> bloc) => bloc.getRequest(
           baseUrl: 'https://jsonplaceholder.typicode.com/',
           handle: 'posts/1',
           fromMap: (json) => PostModel.fromJson(json),
@@ -56,7 +54,7 @@ void main() {
       blocTest(
         'expect assertion if fromMap is neither passed with the function not with cubit',
         build: () => cubitWithoutFromMap,
-        act: (bloc) => expect(
+        act: (RequestCubit<PostModel> bloc) => expect(
           bloc.getRequest(
             baseUrl: 'https://jsonplaceholder.typicode.com/',
             handle: 'posts/1',
@@ -89,7 +87,7 @@ void main() {
       blocTest(
         'pass if baseUrl is given with request function',
         build: () => cubit,
-        act: (bloc) => bloc.getRequest(
+        act: (RequestCubit<PostModel> bloc) => bloc.getRequest(
           baseUrl: 'https://jsonplaceholder.typicode.com/',
           handle: 'posts/1',
         ),
@@ -98,7 +96,7 @@ void main() {
       blocTest(
         'expect assertion if baseUrl is neither given in config nor with request function',
         build: () => cubit,
-        act: (bloc) => expect(
+        act: (RequestCubit<PostModel> bloc) => expect(
           bloc.getRequest(
             handle: 'posts/1',
           ),
@@ -113,7 +111,7 @@ void main() {
 
           return cubit;
         },
-        act: (bloc) => bloc.getRequest(
+        act: (RequestCubit<PostModel> bloc) => bloc.getRequest(
           handle: 'posts/1',
         ),
       );
@@ -142,7 +140,7 @@ void main() {
       blocTest(
         'pass if baseUrl is given with request function',
         build: () => cubit,
-        act: (bloc) => bloc.postRequest(
+        act: (RequestCubit<PostModel> bloc) => bloc.postRequest(
           baseUrl: 'https://jsonplaceholder.typicode.com/',
           handle: 'posts/1',
         ),
@@ -155,7 +153,7 @@ void main() {
 
           return cubit;
         },
-        act: (bloc) => expect(
+        act: (RequestCubit<PostModel> bloc) => expect(
           bloc.postRequest(
             handle: 'posts/1',
           ),
@@ -170,7 +168,7 @@ void main() {
 
           return cubit;
         },
-        act: (bloc) => bloc.postRequest(
+        act: (RequestCubit<PostModel> bloc) => bloc.postRequest(
           handle: 'posts/1',
         ),
       );
